feat: browser-consumable entry (@marko/language-server/browser)
Expose the embedded services and the seams a non-Node host must fill so the
language server can run in a Web Worker, without the vscode-languageserver/node
connection or the jsdom-backed HTML plugin. The Node entry, tests and VS Code
extension are unaffected.
