+++ .changeset/marko-lsp-browser-build.md
@@ -1,0 +1,5 @@
+---
+"@marko/language-server": minor
+---
+
+Add `@marko/language-server/browser`, a browser-consumable entry that exposes the embedded services (Marko + TypeScript + CSS) and the seams a non-Node host must fill, so the language server can run in a Web Worker (eg the markojs.com playground).
