+++ cspell.json
@@ -26,6 +26,7 @@
     "Manthan",
     "marko",
     "Marko",
+    "markojs",
     "mitata",
     "mocharc",
     "nvim",
