+++ packages/language-server/package.json
@@ -20,6 +20,11 @@
     ".": {
       "import": "./dist/index.mjs",
       "default": "./dist/index.js"
+    },
+    "./browser": {
+      "types": "./dist/browser.d.ts",
+      "import": "./dist/browser.mjs",
+      "default": "./dist/browser.js"
     }
   },
   "main": "./dist/index.js",
